OrdinalRegressionViz: Visualize Ordinal Probit Regression Models with 'ggplot2'

Publication-ready 'ggplot2' visualizations for ordinal probit regression models fitted with the 'ordinal' or 'brms' packages, framed in signal detection theory (SDT) terms: receiver operating characteristic (ROC) curves with uncertainty bands, latent-distribution plots with response thresholds, grouped posterior predictive checks, and forest plots for Bayesian random-effects meta-analyses. The visualizations accompany Simchon, Zipori, Teitelbaum, Lewandowsky, and van der Linden (2026) doi:10.1016/j.copsyc.2025.102194 .

See also

Author

Maintainer: Tomer Zipori tomerzip@post.bgu.ac.il

Authors:

Tomer Zipori tomerzip@post.bgu.ac.il